+++ src/legacy/cxgo_server__magic__plyrgide.go
@@ -21,7 +21,7 @@ func nox_xxx_netReportGuideAward_4D8000(a1 int32, a2 int8, a3 int8, a4 int32) in
 		if a4 != 0 {
 			*(*uint8)(unsafe.Add(unsafe.Pointer(&a1), 2)) = uint8(int8(int32(a3) | 0x80))
 		}
-		result = nox_xxx_netSendPacket1_4E5390(int32(*(*uint8)(unsafe.Add(unsafe.Pointer(uintptr(*(*uint32)(unsafe.Add(unsafe.Pointer(uintptr(v5)), 276)))), 2064))), int32(uintptr(unsafe.Pointer(&a1))), 3, 0, 1)
+		result = nox_xxx_netSendPacket1_4E5390(int32(*(*uint8)(unsafe.Add(unsafe.Pointer(uintptr(*(*uint32)(unsafe.Add(unsafe.Pointer(uintptr(v5)), 276)))), 2064))), unsafe.Pointer(&a1), 3, 0, 1)
 	}
 	return result
 }
